requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# OpenAI API for AI features
openai==1.3.0

# Environment variables management
python-dotenv==1.0.0

# Email functionality
Flask-Mail==0.9.1

# Date and time utilities
python-dateutil==2.8.2

# HTTP requests for external APIs
requests==2.31.0

# Data validation
marshmallow==3.20.1

# Payment processing (Stripe)
stripe==7.8.0

# Task scheduling
APScheduler==3.10.4

# File handling
Pillow==10.1.0

# Production server
gunicorn==21.2.0

# Testing (optional for development)
pytest==7.4.3
pytest-flask==1.3.0

